+++ R/features-xcms3.R
@@ -23,7 +23,7 @@ setMethod("delete", "featuresXCMS3", function(obj, i = NULL, j = NULL, ...)
     if (is.null(j) && !setequal(analyses(old), analyses(obj)))
         obj@xdata <- xcms::filterFile(obj@xdata, which(analyses(old) %in% analyses(obj)))
     else if (!is.null(j)) # sync features
-        obj@xdata <- xcms::filterChromPeaks(obj@xdata, getKeptXCMSPeakInds(old, obj))
+        obj@xdata <- xcms::filterChromPeaks(obj@xdata, getKeptXCMSPeakInds(old, obj, obj@xdata))
     return(obj)
 })